ssdb添加lucene检索,添加停止判定机制,停止有待测试,一直阻塞式抽取数据写索引,除非命令停止

+++ java/functions/lucene_ssdb/src/main/java/HttpServer.java
@@ -1,6 +1,9 @@
-import api.AddIndex;
+import api.DelAllIndex;
 import api.NewIndex;
 import api.QueryIndex;
+import api.StartIndex;
+import api.StopIndex;
+import bean.LSException;
 import org.apache.log4j.PropertyConfigurator;
 import util.Constants;
 import org.apache.log4j.Logger;
@@ -11,6 +14,8 @@ import util.Utils;
 import java.lang.management.ManagementFactory;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
+import java.nio.file.LinkOption;
+import java.nio.file.Path;
 
 public class HttpServer {
 
@@ -27,9 +32,11 @@ public class HttpServer {
         httpServer = new Server(Constants.httpPort);
         ServletHandler handler = new ServletHandler();
         httpServer.setHandler(handler);
-        handler.addServletWithMapping(NewIndex.class, "/newIndex");
-        handler.addServletWithMapping(AddIndex.class, "/addIndex");
-        handler.addServletWithMapping(QueryIndex.class, "/queryIndex");
+        handler.addServletWithMapping(NewIndex.class, "/create");
+        handler.addServletWithMapping(QueryIndex.class, "/query");
+        handler.addServletWithMapping(StartIndex.class, "/start");
+        handler.addServletWithMapping(StopIndex.class, "/stop");
+        handler.addServletWithMapping(DelAllIndex.class, "/delAll");
         httpServer.start();
         httpServer.join();
     }
@@ -48,13 +55,17 @@ public class HttpServer {
         Logger logger = Logger.getLogger(HttpServer.class);
         try {
             if ("start".equals(args[0])) {
+                Path pf = Constants.varDir.resolve("pid");
+                if (!Files.notExists(pf, LinkOption.NOFOLLOW_LINKS)) throw new LSException("server pid is exit");
                 HttpServer httpServer = new HttpServer();
                 Runtime.getRuntime().addShutdownHook(new Thread(httpServer::stopHttpServer));
-                httpServer.startHttpServer();
-                Files.createDirectory(Constants.varDir);
                 Files.write(Constants.varDir.resolve("pid"), ManagementFactory.getRuntimeMXBean()
                         .getName().split("@")[0].getBytes(StandardCharsets.UTF_8));
-            } else Utils.stopPid(Constants.varDir.resolve("pid"));
+                httpServer.startHttpServer();
+            } else {
+                int exit = Utils.stopPid(Constants.varDir.resolve("pid"));
+                logger.info("httpServer stop exit:[" + exit + "]");
+            }
         } catch (Exception e) {
             logger.error(e.getCause() == null ? e.getMessage() : e.getCause());
             System.exit(1);
